requirements.txt
streamlit
firebase-admin
pandas
qrcode
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
